+++ server/src/main/java/org/realityforge/replicant/server/transport/ReplicantSessionManager.java
@@ -4,8 +4,8 @@ import javax.annotation.Nonnull;
 import org.realityforge.replicant.server.ChannelDescriptor;
 import org.realityforge.ssf.SessionManager;
 
-public interface ReplicantSessionManager<T extends ReplicantSession>
-  extends SessionManager<T>
+public interface ReplicantSessionManager
+  extends SessionManager<ReplicantSession>
 {
   @Nonnull
   ChannelMetaData getChannelMetaData( @Nonnull ChannelDescriptor descriptor );