+++ server/src/main/java/org/realityforge/replicant/server/transport/ReplicantSessionManagerImpl.java
@@ -13,9 +13,9 @@ import org.realityforge.ssf.InMemorySessionManager;
 /**
  * Base class for session managers.
  */
-public abstract class ReplicantSessionManagerImpl<T extends ReplicantSession>
-  extends InMemorySessionManager<T>
-  implements EntityMessageEndpoint, ReplicantSessionManager<T>
+public abstract class ReplicantSessionManagerImpl
+  extends InMemorySessionManager<ReplicantSession>
+  implements EntityMessageEndpoint, ReplicantSessionManager
 {
   /**
    * Send messages to the specified session.
@@ -27,7 +27,7 @@ public abstract class ReplicantSessionManagerImpl<T extends ReplicantSession>
    * @param changeSet the messages to be sent along to the client.
    * @return the packet created.
    */
-  protected Packet sendPacket( final T session, @Nullable final String etag, @Nonnull final ChangeSet changeSet )
+  protected Packet sendPacket( final ReplicantSession session, @Nullable final String etag, @Nonnull final ChangeSet changeSet )
   {
     final String requestID = (String) getRegistry().getResource( ReplicantContext.REQUEST_ID_KEY );
     getRegistry().putResource( ReplicantContext.REQUEST_COMPLETE_KEY, Boolean.FALSE );
@@ -71,7 +71,7 @@ public abstract class ReplicantSessionManagerImpl<T extends ReplicantSession>
    * @return the packet or null if no packet is ready.
    */
   @Nullable
-  protected Packet pollPacket( @Nonnull final T session, final int lastSequenceAcked )
+  protected Packet pollPacket( @Nonnull final ReplicantSession session, final int lastSequenceAcked )
   {
     final PacketQueue queue = session.getQueue();
     queue.ack( lastSequenceAcked );